Move styles to CSS files.
