+++ src/cards/cover/create.js
@@ -1,6 +1,6 @@
 import { addActions } from "../../tools/tap-actions.js";
 import { createElement, forwardHaptic } from "../../tools/utils.js";
-import styles from "./styles.js";
+import styles from "./styles.css";
 
 export function createStructure(context) {
     context.elements = {};
