+++ src/cards/empty-column/styles.css
@@ -1,0 +1,4 @@
+.empty-column {
+    display: flex;
+    width: 100%;
+}
