--- src/cards/empty-column/styles.js
@@ -1,6 +1,0 @@
-export default `
-    .empty-column {
-        display: flex;
-        width: 100%;
-    }
-`;
