+++ src/cards/pop-up/backdrop.css
@@ -1,0 +1,25 @@
+.bubble-backdrop {
+  position: fixed;
+  background-color: var(--bubble-backdrop-background-color);
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  z-index: 4;
+  opacity: 0;
+  transition: opacity 0.3s;
+  transition-delay: .1s;
+  display: flex;
+  backdrop-filter: var(--custom-backdrop-filter);
+  -webkit-backdrop-filter: var(--custom-backdrop-filter);
+  transform: translate3d(0, 0, 0);
+}
+
+.bubble-backdrop.is-visible {
+  opacity: 1;
+}
+
+.bubble-backdrop.is-hidden {
+  opacity: 0;
+  pointer-events: none;
+}
