+++ src/cards/separator/create.js
@@ -1,6 +1,6 @@
 import { createElement } from "../../tools/utils.js";
 import { addActions } from "../../tools/tap-actions.js";
-import styles from "./styles.js";
+import styles from "./styles.css";
 
 export function createStructure(context) {
   context.elements = {};
