+++ webpack.config.js
@@ -1,6 +1,17 @@
 const webpack = require ('webpack');
 const path = require('path');
 
+const rules = [
+  {
+     test: /\.css/,
+     type: 'asset/source',
+  }
+];
+
+const performance = {
+  hints: false,
+}
+
 module.exports = [
   {
     mode: 'production',
@@ -8,6 +19,10 @@ module.exports = [
       'bubble-card': './src/bubble-card.js',
       'bubble-pop-up-fix': './src/bubble-pop-up-fix.js'
     },
+    module: {
+      rules,
+    },
+    performance,
     output: {
       path: path.resolve(__dirname, 'dist'),
       filename: '[name].js'
@@ -22,6 +37,10 @@ module.exports = [
       'bubble-card': './src/bubble-card.js',
       'bubble-pop-up-fix': './src/bubble-pop-up-fix.js'
     },
+    module: {
+      rules,
+    },
+    performance,
     output: {
       path: path.resolve('/Volumes/config/www'),
       filename: '[name].js'
